Move `ESEFPluginData` to dedicated module
